requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8
python-dotenv>=1.0
vaderSentiment>=3.3.2
openpyxl>=3.1
joblib>=1.4
ipywidgets>=8.1
